fastapi-sr/requirements.txt
fastapi>=0.100.0
uvicorn[standard]
Pillow>=10.0.0
torch>=2.0.0
numpy>=1.20.0
python-multipart
jinja2
